CFP2-96: Move the image-upload component to the forms folder
